--- backend/routes/api/swagger/index.js
@@ -1,15 +1,0 @@
-import express from 'express';
-import path from 'path';
-import swaggerUi from 'swagger-ui-express';
-import YAML from 'yamljs';
-import { fileURLToPath } from 'url';
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-const router = express.Router();
-const swaggerSpec = YAML.load(path.resolve(__dirname,'../../..','swagger/build', 'bundle.yaml'));
-
-router.use('/api-docs', swaggerUi.serve);
-router.get('/api-docs', swaggerUi.setup(swaggerSpec));
-
-export default router;
